Move UseCase to application/usecase
