+++ src/main/java/com/alvarx4z/woja/infrastructure/configuration/UseCaseConfiguration.java
@@ -1,6 +1,6 @@
 package com.alvarx4z.woja.infrastructure.configuration;
 
-import com.alvarx4z.woja.application.GetAllExpansionsUseCase;
+import com.alvarx4z.woja.application.usecase.GetAllExpansionsUseCase;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
